requirements.txt
nltk==3.8.1
scikit-learn==1.2.2
numpy==1.24.3
scipy==1.10.1
joblib==1.2.0
threadpoolctl==3.1.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.0
docx2txt==0.8
pdfminer.six==20221105
